+++ app/components/ui/icon/icon.tsx
@@ -4,7 +4,24 @@ import { tv } from '#/utils/tv';
 
 import iconsHref from './icons.svg';
 
-type IconName = 'user';
+type IconName =
+  | 'calendar'
+  | 'dices'
+  | 'folder'
+  | 'folder-sync'
+  | 'folders'
+  | 'house'
+  | 'log-out'
+  | 'menu'
+  | 'moon'
+  | 'pilcrow'
+  | 'scale'
+  | 'shield'
+  | 'smile-plus'
+  | 'trophy'
+  | 'user'
+  | 'users'
+  | 'x';
 
 const styles = tv({ base: 'inline size-5' });