+++ app/components/ui/index.ts
@@ -1,4 +1,5 @@
 export * from './provider';
 
 export * from './button/button';
+export * from './link/link';
 export * from './menu/menu';